+++ apps/web/src/app/(content)/blog/[id]/[slug]/paragraph.tsx
@@ -1,0 +1,13 @@
+import type { ReactElement, ReactNode } from 'react';
+
+export function Paragraph(props: { children?: ReactNode }) {
+  if (typeof props.children !== 'string') {
+    if (
+      typeof (props.children as ReactElement).type === 'function' ||
+      (props.children as ReactElement).type === 'img'
+    ) {
+      return <>{props.children}</>;
+    }
+  }
+  return <p {...props} />;
+}
