+++ apps/web/src/app/(content)/blog/post-components.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Markdown from 'markdown-to-jsx';
 import Image from 'next/image';
 import Link from 'next/link';
