+++ apps/web/sst-env.d.ts
@@ -1,0 +1,4 @@
+/* tslint:disable */
+/* eslint-disable */
+/// <reference path="../../sst-env.d.ts" />
+export {};
